job_posting/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
anthropic>=0.30.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
numpy>=1.24.0
pydantic>=2.0.0
torch>=2.0.0
transformers>=4.30.0